applied PR feedback
