+++ src/applications/vaos/services/location/index.js
@@ -1,3 +1,5 @@
+import environment from 'platform/utilities/environment';
+
 /*
  * Functions in here should map a var-resources API request to a similar response from
  * a FHIR resource request
@@ -211,4 +213,31 @@ export function getFacilityIdFromLocation(location) {
  */
 export function getSiteIdFromFakeFHIRId(id) {
   return parseId(id).substr(0, 3);
+}
+
+/**
+ * Converts back from a real facility id to our test facility ids
+ * in lower environments
+ *
+ * @export
+ * @param {String} facilityId - facility id to convert
+ * @returns A facility id with either 442 or 552 replaced with 983 or 984
+ */
+export function getTestFacilityId(facilityId) {
+  if (!environment.isProduction() && facilityId) {
+    return facilityId.replace('442', '983').replace('552', '984');
+  }
+
+  return facilityId;
+}
+
+/**
+ * Returns formatted address from facility details object
+ *
+ * @param {*} facility - facility details object
+ */
+export function formatFacilityAddress(facility) {
+  return `${facility.address?.line.join(', ')}, ${facility.address?.city}, ${
+    facility.address?.state
+  } ${facility.address?.postalCode}`;
 }
