+++ src/applications/vaos/services/location/transformers.js
@@ -311,14 +311,3 @@ export function setSupportedSchedulingMethods({
 export function transformFacilities(facilities) {
   return facilities.map(transformFacility);
 }
-
-/**
- * Returns formatted address from facility details object
- *
- * @param {*} facility - facility details object
- */
-export function formatFacilityAddress(facility) {
-  return `${facility.address?.line.join(', ')}, ${facility.address?.city}, ${
-    facility.address?.state
-  } ${facility.address?.postalCode}`;
-}
