+++ src/applications/vaos/tests/services/appointment/transformers.unit.spec.js
@@ -8,10 +8,12 @@ import {
 import {
   transformConfirmedAppointments,
   transformPendingAppointments,
+} from '../../../services/appointment/transformers';
+import {
   CANCELLED_APPOINTMENT_SET,
   PAST_APPOINTMENTS_HIDE_STATUS_SET,
   FUTURE_APPOINTMENTS_HIDE_STATUS_SET,
-} from '../../../services/appointment/transformers';
+} from '../../../services/appointment';
 import { transformATLASLocation } from '../../../services/location/transformers';
 
 const now = moment();
